requirements.txt
annotated-types==0.7.0
anyio==4.5.0
astroid==3.3.3
blinker==1.8.2
build==1.2.2
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dill==0.3.8
distro==1.9.0
easyocr==1.7.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.9.0
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
imageio==2.35.1
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
lazy_loader==0.4
MarkupSafe==2.1.5
mccabe==0.7.0
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==2.1.1
openai==1.47.0
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
pip-tools==7.4.1
platformdirs==4.3.6
pyclipper==1.3.0.post5
pydantic==2.9.2
pydantic_core==2.23.4
pylint==3.3.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyproject_hooks==1.1.0
python-bidi==0.6.0
python-dotenv==1.0.1
PyYAML==6.0.2
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
sniffio==1.3.1
sympy==1.13.3
tifffile==2024.9.20
tomlkit==0.13.2
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
Werkzeug==3.0.4
